Bug fixes, folder will now not be deleted - if there is no folder lel. Folder will be created if there is no folder

+++ src/statistics.py
@@ -1,6 +1,8 @@
 import os
 import shutil
 
+from util import is_directory, make_directories_for_file
+
 
 class Statistics(object):
     """
@@ -13,16 +15,19 @@ class Statistics(object):
 
     def reset(self):
         path = os.path.join("statistics", self.__name)
-        shutil.rmtree(path)
+        if is_directory(path):
+            shutil.rmtree(path)
 
     def save_per_epoch(self, epoch, time, loss, words):
         path = os.path.join("statistics", self.__name, "1_epoch_data.txt")
         words = "\t".join(words)
+        make_directories_for_file(path)
         with open(path, "a+", encoding="utf-8") as fp:
             print(f"{epoch:5d}\t{time:10d}\t{loss:14.10f}\t{words}", file=fp)
 
     def save_per_period(self, epoch, train_acc, test_acc):
         path = os.path.join("statistics", self.__name, "2_period_data.txt")
+        make_directories_for_file(path)
         with open(path, "a+", encoding="utf-8") as fp:
             print(f"{epoch:5d}\t{100*train_acc:9.6f}\t{100*test_acc:9.6f}\t", file=fp)
